% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addalpha.R
\name{addalpha}
\alias{addalpha}
\title{Adds transparency to a list of colors}
\usage{
addalpha(cols, alpha)
}
\arguments{
\item{cols}{integer or chr; if integer from 1-9 takes the value from palette(); if chr a named color from colors()}

\item{alpha}{numeric, specifying the opacity (1) or transparency (0)}
}
\value{
chr vector of colors
}
\description{
Adds transparency to a list of colors
}
\examples{
my_red <- addalpha('red', 0.4)
plot(0,0,col = addalpha('red', 0.1), pch = 16, cex=2)
# more than one color at once
my_col_vec <- addalpha(c(1, 2, 'red', 'navy'), 0.4)
scales::show_col(my_col_vec)
}
