% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balci2010.R
\name{balci2010}
\alias{balci2010}
\title{Peak individual trial analysis using moving average}
\usage{
balci2010(tasa_norm, bines)
}
\arguments{
\item{tasa_norm, }{numeric}

\item{bines, }{numeric}
}
\value{
a numeric vector with start, stop, spread and argmax (the bin at which response rate is max)
}
\description{
Peak individual trial analysis using moving average
}
\details{
Based on Balci et al 2010
}
