% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_fi.r
\name{bp_fi}
\alias{bp_fi}
\title{Single breakpoint algorithm, as the one used in Guilhardi & Church 2004
Works in fixed interval trials.}
\usage{
bp_fi(r_times, if_val)
}
\arguments{
\item{r_times}{numeric, the times at which a response was emitted in a trial}

\item{if_val}{numeric, the duration of the IF interval}
}
\value{
A data frame of 3 columns
bp: a numeric value which corresponds to the time at which a break point was detected
r1: a numeric value of the response rate \emph{before} the breakpoint
r2: a umeric value of the responser rate \emph{after} the breakpoint
}
\description{
Single breakpoint algorithm, as the one used in Guilhardi & Church 2004
Works in fixed interval trials.
}
\details{
This algorithm performs an extensive search of every combination (t1, t2)
where t1 starts in the first response throw (length(r_times) - 1)
}
\examples{
data("r_times")
r_times <- r_times[r_times < 60]
single_bp <- bp_fi(r_times, 60)
par(las = 1)
plot(r_times, seq_along(r_times),
  xlim = c(0, max(r_times)),
  main = "Cummulative Record",
  xlab = "Time (s)",
  ylab = "Cum Resp",
  col = 2, type = "s"
)
abline(v = single_bp$bp)
}
