% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp_km.R
\name{bp_km}
\alias{bp_km}
\title{Break point using kmeans}
\usage{
bp_km(r_times, min_x, max_x, time_res)
}
\arguments{
\item{time_res}{}
}
\value{
data.frame
}
\description{
Break point using kmeans
}
\examples{
data("r_times")
bps_km <- bp_km(r_times, 0, 180, 2)
TODO: Error in dimnames(x) <- dn :
}
