% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curv_indexes.R
\name{curv_index_fry}
\alias{curv_index_fry}
\title{Curvature index using the Fry derivation}
\usage{
curv_index_fry(cr, t, DT, b)
}
\arguments{
\item{cr}{numeric, cumulative response}

\item{t}{numeric, time (or the x axis in a cumulative response plot)}

\item{DT}{numeric, the number of subintervals}

\item{b}{numeric}
}
\value{
the curvature index as exposed by Fry
}
\description{
Curvature index using the Fry derivation
}
