% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curv_indexes.R
\name{curv_index_int}
\alias{curv_index_int}
\title{Curvature index by numerical integration}
\usage{
curv_index_int(cr, t)
}
\arguments{
\item{cr}{numeric, cumulative response}

\item{t}{numeric, time (or the x axis in a cumulative response plot)}

\item{h}{numeric, the height of a rect angule which corresponds to the max cum rate}
}
\value{
a numeric value that is the proportion of a rect triangle area minus
the area under the curve
}
\description{
Curvature index by numerical integration
}
\examples{
data('r_times')
r_times <- r_times[r_times < 60]
cr <- seq_along(r_times)
par(las = 1)
plot(r_times, cr, type = 's')
}
