% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_example.R
\docType{data}
\name{dd_example}
\alias{dd_example}
\title{An example dataset of delays and normalized subjective values}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{delay}{The delay duration (in seconds).}
\item{norm_sv}{The normalized subjective value corresponding to the delay.}
}
}
\usage{
dd_example
}
\description{
A dataset containing delay durations and their respective normalized subjective values (norm_sv).
}
\examples{
data(dd_example)
head(dd_example)
}
\keyword{datasets}
