% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{fi60_raw_from_med}
\alias{fi60_raw_from_med}
\title{Raw Fixed Interval Data}
\format{
A character vector containing lines of data from the file.
}
\source{
The raw data was read from the file: \code{inst/extdata/fi60_raw.txt}.
}
\usage{
fi60_raw_from_med
}
\description{
Raw Fixed Interval Data
}
\details{
A dataset containing raw data from a fixed interval (FI) experiment.

This dataset is obtained from a raw file generated during an FI experiment. It provides raw, unprocessed behavioral data.
}
\keyword{datasets}
