% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gauss_example_2}
\alias{gauss_example_2}
\title{Gaussian Example 2 Data}
\format{
A data frame with 91 rows and 2 columns:
\describe{
\item{Bin}{Numeric. The bin number.}
\item{Response_Average}{Numeric. The average response in each bin.}
}
}
\source{
Generated as part of a synthetic example for the task.
}
\usage{
gauss_example_2
}
\description{
This dataset contains a series of bins and corresponding response averages, this time with a slightly different distribution and experimental task design.
}
\keyword{datasets}
