% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_fit.r
\name{gaussian_fit}
\alias{gaussian_fit}
\title{Fit a gaussian plus a ramp component for the performance in a peak procedure task}
\usage{
gaussian_fit(
  respuestas,
  t,
  par = list(a = 0.1, d = 0.1, t0 = 18, b = 10, c = 1),
  max.iter = 500
)
}
\arguments{
\item{respuestas}{numeric, vector of response or response rate}

\item{t}{numeric, time bins}

\item{par}{a list of parameters for the gaussian + linear;
see Buhusi, C. V., Perera, D., & Meck, W. H. (2005) for an explanation}

\item{max.iter}{numeric, max number of iterations}
}
\value{
a numeric vector of coefficients
}
\description{
Fit a gaussian plus a ramp component for the performance in a peak procedure task
}
\details{
This algorithm uses the nonlinear least squares nls.lm from the minpack.lm package
}
