% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bins.r
\name{get_bins}
\alias{get_bins}
\title{get_bins(x, 0, 20, 5)Returns1  5  5  5  5  5 10 10 10 10 10 15 15 15 15 15 20 20 20 20 20set.seed(10)

x <- runif(20, 0, 10)
get_bins(x, 0, 10, 0.5)Returns1] 5.5 3.5 4.5 7.0 1.0 2.5 3.0 3.0 6.5 4.5 7.0 6.0 1.5 6.0 4.0 4.5 1.0 3.0 4.0 8.5}
\usage{
get_bins(x, x_min, x_max, res)
}
\description{
get_bins(x, 0, 20, 5)Returns1  5  5  5  5  5 10 10 10 10 10 15 15 15 15 15 20 20 20 20 20set.seed(10)

x <- runif(20, 0, 10)
get_bins(x, 0, 10, 0.5)Returns1] 5.5 3.5 4.5 7.0 1.0 2.5 3.0 3.0 6.5 4.5 7.0 6.0 1.5 6.0 4.0 4.5 1.0 3.0 4.0 8.5
}
