% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_overlap.R
\name{hist_over}
\alias{hist_over}
\title{Overlap two histograms}
\usage{
hist_over(
  var1,
  var2,
  name1 = "",
  name2 = "",
  breaks = 30,
  legend = T,
  main0 = "",
  alpha0 = 0.5,
  grey = 0,
  border = T,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Overlap two histograms
}
\examples{
var1 <- rnorm(100)
var2 <- rexp(100)
hist_over(var1, var2, name1 = expression(N(mu == 0, sigma == 1)), name2 = expression(exp(lambda == 1)), grey = T)
}
