% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hyp_data}
\alias{hyp_data}
\title{Simulated Data for Hyperbolic Discounting}
\format{
A list with 3 elements:
\describe{
\item{sv}{A numeric vector of normalized subjective values with added noise.}
\item{delay}{A numeric vector of delays (in seconds).}
\item{real_k}{The real value of the discounting parameter.}
}
}
\source{
Generated using a custom simulation function.
}
\usage{
hyp_data
}
\description{
Simulated Data for Hyperbolic Discounting
}
\details{
A list of simulated data for fitting hyperbolic discounting models.

This dataset was generated to simulate the behavior of a hyperbolic discounting function.
It is commonly used in behavioral economics and psychology to study delay discounting behaviors.
}
\keyword{datasets}
