% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kld_d.r
\name{kld_d}
\alias{kld_d}
\title{Computes the Kullback-Leibler divergence based on kernel density estimates}
\usage{
kld_d(x, y, from_a, to_b)
}
\arguments{
\item{x}{numeric, the values from a sample p}

\item{y}{numeric, the values from a sample q}

\item{from_a}{numeric, the lower limit of the integration}

\item{to_b}{numeric, the upper limit of the integration}
}
\value{
a numeric value that is the kl divergence
}
\description{
Computes the Kullback-Leibler divergence based on kernel density estimates
}
\examples{
set.seed(123)
p <- rnorm(100)
q <- rnorm(100)
kld_d(p, q, -Inf, Inf) # 0.07579204
q <- rnorm(100, 10, 4)
kld_d(p, q, -Inf, Inf) # 7.769912
}
