% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{r_times}
\alias{r_times}
\title{Reaction Times from Peak Procedure}
\format{
A numeric vector with 132 elements representing reaction times.
}
\source{
Generated during a behavioral analysis experiment.
}
\usage{
r_times
}
\description{
Reaction Times from Peak Procedure
}
\details{
A dataset containing reaction times (in seconds) from an experiment using the peak procedure.

These times are derived from a peak procedure experiment, typically used in behavioral experiments to measure timing abilities in subjects.
}
\keyword{datasets}
