% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_med.r
\name{read_med}
\alias{read_med}
\title{Process MED to csv based on standard data structure event.time}
\usage{
read_med(fname, save_file = FALSE, path_save = NULL, col_r = "C:", out = T)
}
\arguments{
\item{fname}{chr, the name of a MED file to read; can include the path directory}

\item{save_file}{logical, Save csv? TRUE or FALSE (default)}

\item{path_save}{chr, Path to folder to save csv files;}

\item{col_r}{chr, Variable of MED-raw to read (a event.time variable)}

\item{out}{logical, output the dataframe processed?}
}
\description{
Process MED to csv based on standard data structure event.time
}
