% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit_normalization.R
\name{unit_normalization}
\alias{unit_normalization}
\title{Min-max normalization (also feature rescaling)}
\usage{
unit_normalization(x)
}
\arguments{
\item{x}{numeric, vector of values to rescale}
}
\value{
A numeric vector rescaled in the range \eqn{x' \in [0, 1]}
}
\description{
Min-max normalization (also feature rescaling)
}
\examples{
x <- 5:100
x_scaled <- unit_normalization(x)
x_scaled
}
