% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bins.r
\docType{data}
\name{x}
\alias{x}
\title{A function to binarize a numeric vector with a given resolution}
\format{
An object of class \code{integer} of length 20.
}
\usage{
x
}
\arguments{
\item{x}{numeric, the vector to be binarized}

\item{x_min}{numeric, the min value of a vector to create the bins (e.g., 0)}

\item{x_max}{numeric, the maximum value of the vector x to binarize}

\item{res}{numeric, the resolution; if x is time, res can be 1 s}
}
\value{
the vector of bins for which x is in
}
\description{
A function to binarize a numeric vector with a given resolution
}
\keyword{datasets}
