% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YEAB-package.R
\docType{package}
\name{YEAB-package}
\alias{YEAB}
\alias{YEAB-package}
\title{YEAB: Analyze Data from Analysis of Behavior Experiments}
\description{
Analyze data from behavioral experiments conducted using MED-PC software developed by Med Associates Inc. Includes functions to fit exponential and hyperbolic models for delay discounting tasks, exponential mixtures for inter-response times, and Gaussian plus ramp models for peak procedure data, among others. For more details, refer to Alcalá et al. (2023) \doi{10.31234/osf.io/8aq2j}.
}
\author{
\strong{Maintainer}: Emmanuel Alcala \email{jealcalat@gmail.com}

Authors:
\itemize{
  \item Rodrigo Sosa \email{rsosas@up.edu.mx}
  \item Victor Reyes \email{zickvic@hotmail.com}
}

}
\keyword{internal}
