% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curv_indexes.R
\name{curv_index_fry}
\alias{curv_index_fry}
\title{Curvature index using the Fry derivation}
\usage{
curv_index_fry(cr, time_in, n, b, if_val)
}
\arguments{
\item{cr}{numeric, cumulative response}

\item{n}{numeric, the number of subintervals}

\item{b}{numeric, the base of the triangle (e.g., FI value)}

\item{t}{numeric, time (or the x axis in a cumulative response plot)}
}
\value{
the curvature index as exposed by Fry
}
\description{
Curvature index using the Fry derivation
}
