% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy_kde2d.R
\name{entropy_kde2d}
\alias{entropy_kde2d}
\title{Shannon entropy in two dimensions}
\usage{
entropy_kde2d(x, y, n_grid = 150)
}
\arguments{
\item{x}{numeric, random vector}

\item{y}{numeric, random vector}

\item{n_grid}{numeric, number of grid cells to evaluate density}
}
\description{
Shannon entropy in two dimensions
}
\examples{
set.seed(123)
x <- rnorm(1000, 0, 1)
set.seed(12)
y <- rnorm(1000, 0, 5)
cov_matr <- cov(cbind(x, y))
sigmas <- diag(cov_matr)
det_sig <- prod(sigmas)
# equivalentemente: det_sig <- det(cov(cbind(x,y)))
# de acuerdo con https://en.wikipedia.org/wiki/Multivariate_normal_distribution#'Differential_entropy
normal_entropy <- function(k, pi, det_sig) {
  # todo lo de la izquierda es una constante; en Maei y Frankland se elimina
  (k / 2) * (1 + log(2 * pi)) + (1 / 2) * log(det_sig)
}
entropia <- normal_entropy(k = 2, pi = pi, det_sig)
entropia # retorna 7.415341
entropy_kde2d(x, y, n_grid = 50)
# retorna 7.445117
# muy cercanos
}
