% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_bout.R
\name{estimate_bout}
\alias{estimate_bout}
\title{Biexponential model for the estimation of whithin and between-bouts}
\usage{
estimate_bout(irt)
}
\arguments{
\item{irt}{numeric, the interresponse times}
}
\value{
numeric vector of the biexponential model parameters
}
\description{
Biexponential model for the estimation of whithin and between-bouts
}
