% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_extractor.R
\name{event_extractor}
\alias{event_extractor}
\title{Slice from a data frame based on starting and ending of a event}
\usage{
event_extractor(dframe, ev0, ev1, evname)
}
\arguments{
\item{dframe}{data frame with events ev0 and ev1 (e.g., start of trial and reinforcement delivery)}

\item{ev0}{event ID start (where the event we want to extract begins)}

\item{ev1}{event ID stop. This event won't be returned, so keep in mind that}

\item{evname}{a string for the event name, for identification purposes. For example
if the event we want to extract is component 1 in a multiple-2 schedule,
this can be eventname = "c1", so when we extract the second
component we can row-combine both in a unique dataframe.}
}
\value{
data frame with nrows x 4 columns of time, events, cum_id and evname
}
\description{
Slice from a data frame based on starting and ending of a event
}
\details{
Works by trials
}
\examples{
# How to use ----
 # # If we have a component starting with 5 and ending with 2 and a dataframe "df"
 # we can extract the data of component "comp52" following the next steps:
 # 0 - From the output of read_med.R function, load the csv file and assing to df
 # 1 - source the event_extractor.R function
 # 2 - use it with the appropiate arguments as follows
component52df <- event_extractor(dframe = df, # enter the data as the 1st arg
                                 ev0 = 5, ev1 = 2, # anter start and stop
                                 evname = "comp52") # enter the event name
}
